Refactors new exception following feedback
