+++ python/semantic_kernel/connectors/ai/open_ai/exceptions/__init__.py
@@ -1,0 +1,1 @@
+# Copyright (c) Microsoft. All rights reserved.
