+++ python/semantic_kernel/connectors/ai/open_ai/services/open_ai_handler.py
@@ -16,7 +16,7 @@ from semantic_kernel.connectors.ai.chat_request_settings import ChatRequestSetti
 from semantic_kernel.connectors.ai.complete_request_settings import (
     CompleteRequestSettings,
 )
-from semantic_kernel.connectors.ai.content_filter_ai_exception import (
+from semantic_kernel.connectors.ai.open_ai.exceptions.content_filter_ai_exception import (
     ContentFilterAIException,
 )
 from semantic_kernel.connectors.ai.open_ai.services.open_ai_model_types import (
@@ -79,22 +79,10 @@ class OpenAIHandler(AIServiceClientBase, ABC):
             )
         except BadRequestError as ex:
             if ex.code == "content_filter":
-                inner_error = ex.body.get("innererror", {})
-                content_filter_code = inner_error.get("code", "")
-                content_filter_results = inner_error.get("content_filter_result", {})
                 raise ContentFilterAIException(
-                    f"{type(self)} service failed to complete the prompt",
-                    ex.param,
-                    ContentFilterAIException.ContentFilterCodes(content_filter_code),
-                    dict(
-                        [
-                            key,
-                            ContentFilterAIException.ContentFilterResult.from_inner_error_result(
-                                values
-                            ),
-                        ]
-                        for key, values in content_filter_results.items()
-                    ),
+                    AIException.ErrorCodes.BadContentError,
+                    f"{type(self)} service encountered a content error",
+                    ex,
                 )
             raise AIException(
                 AIException.ErrorCodes.ServiceError,
